+++ src/test/java/org/seedstack/oauth/ImplicitFlowIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2013-2020, The SeedStack authors <http://seedstack.org>
+ * Copyright © 2013-2021, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,6 +8,7 @@
 package org.seedstack.oauth;
 
 import io.restassured.http.Header;
+import io.restassured.response.ResponseBody;
 import org.assertj.core.util.Lists;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -22,7 +23,6 @@ import static io.restassured.RestAssured.given;
 
 @RunWith(SeedITRunner.class)
 @LaunchWithUndertow
-@Ignore
 public class ImplicitFlowIT {
     @Configuration("runtime.web.baseUrl")
     private String baseUrl;
@@ -35,11 +35,40 @@ public class ImplicitFlowIT {
         TokenData token = tokenBuilder.buildToken("test", Lists.newArrayList("email", "profile"));
         given()
                 .log().uri()
-                .header(new Header("Authorization", token.getAccess_token()))
+                .header(new Header("Authorization", "Bearer " + token.getAccess_token()))
                 .expect()
                 .log().status()
                 .statusCode(200)
                 .when()
                 .get(baseUrl + "/api/profile");
+    }
+
+    @Test
+    public void requestWithoutTokenShould401() {
+        TokenBuilder tokenBuilder = new TokenBuilder(oauthConfig);
+        TokenData token = tokenBuilder.buildToken("test", Lists.newArrayList("email", "profile"));
+        given()
+                .log().uri()
+                .expect()
+                .log().status()
+                .statusCode(401)
+                .when()
+                .get(baseUrl + "/api/profile");
+    }
+
+    @Test
+    public void requestWithInvalidTokenShould401() {
+        TokenBuilder tokenBuilder = new TokenBuilder(oauthConfig);
+        TokenData token = tokenBuilder.buildToken("test", Lists.newArrayList("email", "profile"));
+        ResponseBody body = given()
+                .log().uri()
+                .header(new Header("Authorization", "Bearer dummy"))
+                .expect()
+                .log().status()
+                .statusCode(401)
+                .when()
+                .get(baseUrl + "/api/profile")
+                .getBody();
+        System.out.println(body.asString());
     }
 }