+++ src/test/java/org/seedstack/oauth/fixtures/provider/TokenResource.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2013-2020, The SeedStack authors <http://seedstack.org>
+ * Copyright © 2013-2021, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,14 +7,20 @@
  */
 package org.seedstack.oauth.fixtures.provider;
 
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
 import org.seedstack.oauth.OAuthConfig;
 import org.seedstack.oauth.fixtures.TokenBuilder;
 import org.seedstack.seed.Configuration;
+
+import javax.ws.rs.FormParam;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Request;
+import javax.ws.rs.core.Response;
+import java.util.ArrayList;
+import java.util.Arrays;
 
 @Path("/provider/create-token")
 public class TokenResource {
@@ -31,16 +37,16 @@ public class TokenResource {
 
     @POST
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createToken() {
-        return Response.ok(tokenData(AuthorizationResource.nonce)).build();
+    public Response createToken(@FormParam("scope") String scope) {
+        return Response.ok(tokenData(AuthorizationResource.nonce, scope)).build();
     }
 
-    private TokenData tokenData(String nonce) {
+    private TokenData tokenData(String nonce, String scope) {
         TokenBuilder tb = new TokenBuilder(oauthConfig);
         tb.setTestInvalidNonce(testInvalidNonce);
         tb.setTestInvalidAudience(testInvalidAudience);
         tb.setTestTokenExpiry(testTokenExpiry);
         tb.setFlagForAccessToken(buildOnlyAccessToken);
-        return tb.buildToken(nonce, oauthConfig.getScopes());
+        return tb.buildToken(nonce, scope == null ? new ArrayList<>() : Arrays.asList(scope.split(" ").clone()));
     }
 }